+++ core/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(SOURCES
-    src/arp_poisoning/arp_poisoning.cpp
-    src/arp_poisoning/create_packet.cpp
+    src/arp/arp_poisoning.cpp
+    src/arp/create_packet.cpp
     src/network_scout/network_scout.cpp
     src/network_scout/sniffing.cpp
     src/common/pcap_to_common.cpp
